+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## master
 
+* Bug Fixes
+  * Fix usage of deprecated modules
+
 ## 2.0.2 (2016-12-08)
 
 * Bug Fixes